Adding the ability to have an optional ice melt threshold failure criterion, which only applies if the material has a porosity> 0.99 (this is our way, for now, of tagging a material as ice wedge)
The change is made by adding the paramter 'ACE Ice Melt Threshold: 1.0e-03' to materials_mechanical_denudation.yaml for the material you wish to have this criterion applied to

Failure by melt is achieved if the number of integration points in an element marked as having failed (having ice saturation below the user-provided melt threshold) exceeds the failure threshold (assumed to be 8 integration points).

This failure criterion is evalauted separately from the 5 mechanical criteria, and so failure of an element is defined by either (failure by mechanical criteria) OR (failure by ice melt).